Fix test, rename
