+++ src/_reducers/trades/TradesProposalInfoReducer.js
