+++ src/_reducers/trades/TradesPurchaseInfoReducer.js
