+++ src/_reducers/trades/TradesUIStateReducer.js
